requirements.txt
office365-rest-python-client
pandas
openpyxl
joblib